% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fullread.gene2acc}
\alias{fullread.gene2acc}
\title{utility for import of gene2accession}
\usage{
fullread.gene2acc()
}
\value{
data.frame with all human gene2go records
}
\description{
utility for import of gene2accession
}
