% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkgoeg.R
\name{fullread.gene2go}
\alias{fullread.gene2go}
\title{utility for import of gene2go}
\usage{
fullread.gene2go()
}
\value{
data.frame with all human gene2go records
}
\description{
utility for import of gene2go
}
