% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fullread.ncbitxt}
\alias{fullread.ncbitxt}
\title{generalized utility for import of NCBI text}
\usage{
fullread.ncbitxt(stub = "gene_info")
}
\value{
data.frame with all human gene2go records
}
\description{
generalized utility for import of NCBI text
}
