% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keys,hsParqEnv-method}
\alias{keys,hsParqEnv-method}
\title{emulate keys method of AnnotationDbi but do not import generic}
\usage{
\S4method{keys}{hsParqEnv}(x, keytype, ...)
}
\arguments{
\item{x}{instance of hsParqEnv}

\item{keytype}{not used}
}
\description{
emulate keys method of AnnotationDbi but do not import generic
}
\examples{
head(keys(org.Hs.egMAP2EG))
}
