% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\docType{data}
\name{org.Hs.egALIAS2EG}
\alias{org.Hs.egALIAS2EG}
\title{self-describing object for ALIAS2EG}
\format{
An object of class \code{hsParqEnv} (inherits from \code{environment}) of length 69601.
}
\usage{
org.Hs.egALIAS2EG
}
\description{
self-describing object for ALIAS2EG
}
\keyword{datasets}
