% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\name{org.Hs.egENSEMBL}
\alias{org.Hs.egENSEMBL}
\title{self-describing object for ENSEMBL}
\usage{
org.Hs.egENSEMBL
}
\description{
self-describing object for ENSEMBL
}
\examples{
org.Hs.egENSEMBL
toTable(org.Hs.egENSEMBL) |> head()
}
