% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\name{org.Hs.egENSEMBLPROT}
\alias{org.Hs.egENSEMBLPROT}
\title{self-describing object for ENSEMBLPROT}
\usage{
org.Hs.egENSEMBLPROT
}
\description{
self-describing object for ENSEMBLPROT
}
\examples{
org.Hs.egENSEMBLPROT
toTable(org.Hs.egENSEMBLPROT) |> head()
}
