% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\name{org.Hs.egENSEMBLPROT2EG}
\alias{org.Hs.egENSEMBLPROT2EG}
\title{self-describing object for ENSEMBLPROT2EG}
\usage{
org.Hs.egENSEMBLPROT2EG
}
\description{
self-describing object for ENSEMBLPROT2EG
}
\examples{
org.Hs.egENSEMBLPROT2EG
toTable(org.Hs.egENSEMBLPROT2EG) |> head()
}
