% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\name{org.Hs.egENSEMBLTRANS}
\alias{org.Hs.egENSEMBLTRANS}
\title{self-describing object for ENSEMBLTRANS}
\usage{
org.Hs.egENSEMBLTRANS
}
\description{
self-describing object for ENSEMBLTRANS
}
\examples{
org.Hs.egENSEMBLTRANS
toTable(org.Hs.egENSEMBLTRANS) |> head()
}
