% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\name{org.Hs.egENSEMBLTRANS2EG}
\alias{org.Hs.egENSEMBLTRANS2EG}
\title{self-describing object for ENSEMBLTRANS2EG}
\usage{
org.Hs.egENSEMBLTRANS2EG
}
\description{
self-describing object for ENSEMBLTRANS2EG
}
\examples{
org.Hs.egENSEMBLTRANS2EG
toTable(org.Hs.egENSEMBLTRANS2EG) |> head()
}
