% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\name{org.Hs.egGENENAME}
\alias{org.Hs.egGENENAME}
\title{self-describing object for GENENAME}
\usage{
org.Hs.egGENENAME
}
\description{
self-describing object for GENENAME
}
\examples{
org.Hs.egGENENAME
toTable(org.Hs.egGENENAME) |> head()
}
