% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\name{org.Hs.egGENETYPE}
\alias{org.Hs.egGENETYPE}
\title{self-describing object for GENETYPE}
\usage{
org.Hs.egGENETYPE
}
\description{
self-describing object for GENETYPE
}
\examples{
org.Hs.egGENETYPE
toTable(org.Hs.egGENETYPE) |> head()
}
