% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\docType{data}
\name{org.Hs.egGO}
\alias{org.Hs.egGO}
\title{self-describing object for GO interface}
\format{
An object of class \code{hsParqEnv} (inherits from \code{environment}) of length 20826.
}
\usage{
org.Hs.egGO
}
\description{
self-describing object for GO interface
}
\keyword{datasets}
