% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\docType{data}
\name{org.Hs.egGO2EG}
\alias{org.Hs.egGO2EG}
\title{self-describing object for GO2EG interface}
\format{
An object of class \code{hsParqEnv} (inherits from \code{environment}) of length 17947.
}
\usage{
org.Hs.egGO2EG
}
\description{
self-describing object for GO2EG interface
}
\keyword{datasets}
