% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\docType{data}
\name{org.Hs.egMAP}
\alias{org.Hs.egMAP}
\title{self-describing object for MAP}
\format{
An object of class \code{hsParqEnv} (inherits from \code{environment}) of length 193790.
}
\usage{
org.Hs.egMAP
}
\description{
self-describing object for MAP
}
\examples{
org.Hs.egMAP
toTable(org.Hs.egMAP) |> head()
}
\keyword{datasets}
