% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\name{org.Hs.egMAP2EG}
\alias{org.Hs.egMAP2EG}
\title{self-describing object for MAP}
\usage{
org.Hs.egMAP2EG
}
\description{
self-describing object for MAP
}
\examples{
org.Hs.egMAP2EG
toTable(org.Hs.egMAP2EG) |> head()
}
