% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\name{org.Hs.egOMIM}
\alias{org.Hs.egOMIM}
\title{self-describing object for OMIM}
\usage{
org.Hs.egOMIM
}
\description{
self-describing object for OMIM
}
\examples{
org.Hs.egOMIM
toTable(org.Hs.egOMIM) |> head()
}
