% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\name{org.Hs.egOMIM2EG}
\alias{org.Hs.egOMIM2EG}
\title{self-describing object for OMIM2EG}
\usage{
org.Hs.egOMIM2EG
}
\description{
self-describing object for OMIM2EG
}
\examples{
org.Hs.egOMIM2EG
toTable(org.Hs.egOMIM2EG) |> head()
}
