% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\name{org.Hs.egSYMBOL}
\alias{org.Hs.egSYMBOL}
\title{self-describing object for SYMBOL}
\usage{
org.Hs.egSYMBOL
}
\description{
self-describing object for SYMBOL
}
\examples{
org.Hs.egSYMBOL
toTable(org.Hs.egSYMBOL) |> head()
}
