% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\name{org.Hs.egSYMBOL2EG}
\alias{org.Hs.egSYMBOL2EG}
\title{self-describing object for SYMBOL2EG}
\usage{
org.Hs.egSYMBOL2EG
}
\description{
self-describing object for SYMBOL2EG
}
\examples{
org.Hs.egSYMBOL2EG
toTable(org.Hs.egSYMBOL2EG) |> head()
}
