% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgObj.R
\name{show,OrgParq-method}
\alias{show,OrgParq-method}
\title{display for parquet representation of GO}
\usage{
\S4method{show}{OrgParq}(object)
}
\arguments{
\item{object}{instance of OrgParq}
}
\description{
display for parquet representation of GO
}
