% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\name{simpleEnvBuilder}
\alias{simpleEnvBuilder}
\title{closure producer for use in zzz.R}
\usage{
simpleEnvBuilder(keys, values, colname)
}
\arguments{
\item{keys}{character vector}

\item{values}{list}

\item{colname}{character of length 1 to be used as the values column name}
}
\description{
closure producer for use in zzz.R
}
