% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\docType{data}
\name{org.Hs.egENSEMBL}
\alias{org.Hs.egENSEMBL}
\title{self-describing object for ENSEMBL}
\format{
An object of class \code{hsParqEnv} (inherits from \code{environment}) of length 38516.
}
\usage{
org.Hs.egENSEMBL
}
\description{
self-describing object for ENSEMBL
}
\keyword{datasets}
