% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\docType{data}
\name{org.Hs.egENSEMBL2EG}
\alias{org.Hs.egENSEMBL2EG}
\title{self-describing object for ENSEMBL2EG}
\format{
An object of class \code{hsParqEnv} (inherits from \code{environment}) of length 63890.
}
\usage{
org.Hs.egENSEMBL2EG
}
\description{
self-describing object for ENSEMBL2EG
}
\examples{
org.Hs.egENSEMBL2EG
toTable(org.Hs.egENSEMBL2EG) |> head()
}
\keyword{datasets}
