% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\docType{data}
\name{org.Hs.egENSEMBLPROT}
\alias{org.Hs.egENSEMBLPROT}
\title{self-describing object for ENSEMBLPROT}
\format{
An object of class \code{hsParqEnv} (inherits from \code{environment}) of length 38516.
}
\usage{
org.Hs.egENSEMBLPROT
}
\description{
self-describing object for ENSEMBLPROT
}
\examples{
org.Hs.egENSEMBLPROT
toTable(org.Hs.egENSEMBLPROT) |> head()
}
\keyword{datasets}
