% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\docType{data}
\name{org.Hs.egENSEMBLTRANS}
\alias{org.Hs.egENSEMBLTRANS}
\title{self-describing object for ENSEMBLTRANS}
\format{
An object of class \code{hsParqEnv} (inherits from \code{environment}) of length 38516.
}
\usage{
org.Hs.egENSEMBLTRANS
}
\description{
self-describing object for ENSEMBLTRANS
}
\examples{
org.Hs.egENSEMBLTRANS
toTable(org.Hs.egENSEMBLTRANS) |> head()
}
\keyword{datasets}
