% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\docType{data}
\name{org.Hs.egENSEMBLTRANS2EG}
\alias{org.Hs.egENSEMBLTRANS2EG}
\title{self-describing object for ENSEMBLTRANS2EG}
\format{
An object of class \code{hsParqEnv} (inherits from \code{environment}) of length 78869.
}
\usage{
org.Hs.egENSEMBLTRANS2EG
}
\description{
self-describing object for ENSEMBLTRANS2EG
}
\examples{
org.Hs.egENSEMBLTRANS2EG
toTable(org.Hs.egENSEMBLTRANS2EG) |> head()
}
\keyword{datasets}
