% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\docType{data}
\name{org.Hs.egGENETYPE}
\alias{org.Hs.egGENETYPE}
\title{self-describing object for GENETYPE}
\format{
An object of class \code{hsParqEnv} (inherits from \code{environment}) of length 193790.
}
\usage{
org.Hs.egGENETYPE
}
\description{
self-describing object for GENETYPE
}
\examples{
org.Hs.egGENETYPE
toTable(org.Hs.egGENETYPE) |> head()
}
\keyword{datasets}
