% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\docType{data}
\name{org.Hs.egMAP2EG}
\alias{org.Hs.egMAP2EG}
\title{self-describing object for MAP}
\format{
An object of class \code{hsParqEnv} (inherits from \code{environment}) of length 2020.
}
\usage{
org.Hs.egMAP2EG
}
\description{
self-describing object for MAP
}
\examples{
org.Hs.egMAP2EG
toTable(org.Hs.egMAP2EG) |> head()
}
\keyword{datasets}
