% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfunc.R
\name{simpleEnvBuilder}
\alias{simpleEnvBuilder}
\title{closure producer for use in zzz.R}
\usage{
simpleEnvBuilder(keys, values, keycolname, valcolname)
}
\arguments{
\item{keys}{character vector}

\item{values}{list}

\item{keycolname}{character of length 1 to be used as the key column name for toTable}

\item{valcolname}{character of length 1 to be used as the values column name for toTable}
}
\description{
closure producer for use in zzz.R
}
