% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toTable,hsParqEnv-method}
\alias{toTable,hsParqEnv-method}
\title{conversion utility}
\usage{
\S4method{toTable}{hsParqEnv}(x, ...)
}
\arguments{
\item{x}{hsParqEnv instance}

\item{\dots}{not used}
}
\description{
conversion utility
}
\examples{
toTable(org.Hs.egGO) |> head(10)
}
